Created a JUnit Test Case for the AddEditNewsStoryView class.
This was used to debug and test the class, which should be finished now,
but I'm keeping it around just in case. We'll have to remember to delete
it. Most of the changes were minor bug fixes, but I also added a Box and
verticalGlue structure to the View, and it looks v nice mwah.